+++ src/eca/messenger.clj
@@ -7,5 +7,6 @@
 (defprotocol IMessenger
   (chat-content-received [this data])
   (tool-server-updated [this params])
+  (config-updated [this params])
   (showMessage [this msg])
   (editor-diagnostics [this uri]))