+++ src/eca/server.clj
@@ -98,6 +98,9 @@
   (chat-content-received [_this content]
     (lsp.server/discarding-stdout
      (lsp.server/send-notification server "chat/contentReceived" content)))
+  (config-updated [_this params]
+    (lsp.server/discarding-stdout
+     (lsp.server/send-notification server "config/updated" params)))
   (tool-server-updated [_this params]
     (lsp.server/discarding-stdout
      (lsp.server/send-notification server "tool/serverUpdated" params)))