+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- New server notification `config/updated` used to notify clients when a relevant config changed (behaviors, models etc).
+- Deprecate info inside `initialize` response, clients should use `config/updated` now.
+
 ## 0.41.0
 
 - Improve anthropic extraPayload requirement when adding models.